Download binaries script refactoring
