+++ js/package.json
@@ -15,14 +15,15 @@
   ],
   "scripts": {
     "postinstall": "npm run install_runtime",
-    "download_runtime": "node ./scripts/download_runtime.js",
+    "download_runtime": "node ./scripts/download-runtime.js",
     "install_runtime": "npm run download_runtime -- --ignore-if-exists",
     "test": "node --test ./tests/*.test.js"
   },
   "binary": {
+    "version": "2024.4.0.0",
     "module_path": "./bin/",
-    "remote_path": "./repositories/openvino_tokenizers/packages/{version}.0/",
-    "package_name": "openvino_tokenizers_{platform}_{version}.0_{arch}.tar.gz",
+    "remote_path": "./repositories/openvino_tokenizers/packages/{version}/",
+    "package_name": "openvino_tokenizers_{platform}_{version}_{arch}.{extension}",
     "host": "https://storage.openvinotoolkit.org"
   },
   "keywords": [
@@ -31,6 +32,7 @@
     "OpenVINO tokenizers"
   ],
   "dependencies": {
+    "adm-zip": "^0.5.16",
     "gunzip-maybe": "^1.4.2",
     "https-proxy-agent": "^7.0.5",
     "tar-fs": "^3.0.6"
